Move dining models to shared folder
