--- PennMobile/General/KeychainAccessible.swift
@@ -1,114 +1,0 @@
-//
-//  KeychainAccessible.swift
-//  PennMobile
-//
-//  Created by Josh Doman on 1/1/20.
-//  Copyright © 2020 PennLabs. All rights reserved.
-//
-
-import Foundation
-
-class KeychainAccessible {
-
-    static let instance = KeychainAccessible()
-
-    private var pennkeyKeychainKey: String {
-        "PennKey"
-    }
-
-    private var passwordKeychainKey: String {
-        "PennKey Password"
-    }
-
-    private var pacCodeKeychainKey: String {
-        "PAC Code"
-    }
-
-    private var diningTokenKeychainKey: String {
-        "Dining Token"
-    }
-
-    func getPennKey() -> String? {
-        let secureStore = getWebLoginSecureStore()
-        do {
-            return try secureStore.getValue(for: pennkeyKeychainKey)
-        } catch {
-            return nil
-        }
-    }
-
-    func getPassword() -> String? {
-        let secureStore = getWebLoginSecureStore()
-        do {
-            return try secureStore.getValue(for: passwordKeychainKey)
-        } catch {
-            return nil
-        }
-    }
-
-    func getPacCode() -> String? {
-        let secureStore = getWebLoginSecureStore()
-        do {
-            return try secureStore.getValue(for: pacCodeKeychainKey)
-        } catch {
-            return nil
-        }
-    }
-
-    func savePennKey(_ pennkey: String) {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.setValue(pennkey, for: pennkeyKeychainKey)
-    }
-
-    func removePennKey() {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.removeValue(for: pennkeyKeychainKey)
-    }
-
-    func savePassword(_ password: String) {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.setValue(password, for: passwordKeychainKey)
-    }
-
-    func removePassword() {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.removeValue(for: passwordKeychainKey)
-    }
-
-    func savePacCode(_ pacCode: String) {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.setValue(pacCode, for: pacCodeKeychainKey)
-    }
-
-    func removePacCode() {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.removeValue(for: pacCodeKeychainKey)
-    }
-
-    private func getWebLoginSecureStore() -> SecureStore {
-        let genericPwdQueryable = GenericPasswordQueryable(service: "PennWebLogin")
-        let secureStore = SecureStore(secureStoreQueryable: genericPwdQueryable)
-        return secureStore
-    }
-
-    func saveDiningToken(_ diningToken: String) {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.setValue(diningToken, for: diningTokenKeychainKey)
-    }
-
-    func getDiningToken() -> String? {
-        let secureStore = getWebLoginSecureStore()
-        do {
-            return try secureStore.getValue(for: diningTokenKeychainKey)
-        } catch {
-            return nil
-        }
-    }
-
-    func removeDiningToken() {
-        let secureStore = getWebLoginSecureStore()
-        try? secureStore.removeValue(for: diningTokenKeychainKey)
-        UserDefaults.standard.clearDiningBalance()
-    }
-
-}
