+++ PennMobile/General/Networking + Analytics/Networking.swift
@@ -19,16 +19,6 @@ public enum Method {
     case other(method: String)
 }
 
-enum NetworkingError: String, Error {
-    case noInternet
-    case parsingError
-    case serverError
-    case jsonError = "JSON error"
-    case authenticationError = "Unable to authenticate"
-    case other
-    var localizedDescription: String { self.rawValue }
-}
-
 extension Method {
     public init(_ rawValue: String) {
         let method = rawValue.uppercased()
